seperated existing application into subapps
